+++ nautilus_trader/adapters/_template/execution.py
@@ -22,6 +22,7 @@ from nautilus_trader.execution.messages import GenerateOrderStatusReport
 from nautilus_trader.execution.messages import GenerateOrderStatusReports
 from nautilus_trader.execution.messages import GeneratePositionStatusReports
 from nautilus_trader.execution.messages import ModifyOrder
+from nautilus_trader.execution.messages import QueryAccount
 from nautilus_trader.execution.messages import QueryOrder
 from nautilus_trader.execution.messages import SubmitOrder
 from nautilus_trader.execution.messages import SubmitOrderList
@@ -55,11 +56,13 @@ class TemplateLiveExecutionClient(LiveExecutionClient):
     | dispose                                    | optional    |
     +--------------------------------------------+-------------+
     | _submit_order                              | required    |
-    | _submit_order_list                         | required    |
+    | _submit_order_list                         | optional    |
     | _modify_order                              | optional    |
     | _cancel_order                              | required    |
     | _cancel_all_orders                         | required    |
     | _batch_cancel_orders                       | optional    |
+    | _query_account                             | optional    |
+    | _query_order                               | optional    |
     | generate_order_status_report               | required    |
     | generate_order_status_reports              | required    |
     | generate_fill_reports                      | required    |
@@ -152,6 +155,11 @@ class TemplateLiveExecutionClient(LiveExecutionClient):
     async def _batch_cancel_orders(self, command: BatchCancelOrders) -> None:
         raise NotImplementedError(
             "method `_batch_cancel_orders` must be implemented in the subclass",
+        )  # pragma: no cover
+
+    async def _query_account(self, command: QueryAccount) -> None:
+        raise NotImplementedError(
+            "method `_query_account` must be implemented in the subclass",
         )  # pragma: no cover
 
     async def _query_order(self, command: QueryOrder) -> None: