+++ nautilus_trader/execution/client.pyx
@@ -33,6 +33,8 @@ from nautilus_trader.core.uuid cimport UUID4
 from nautilus_trader.execution.messages cimport CancelAllOrders
 from nautilus_trader.execution.messages cimport CancelOrder
 from nautilus_trader.execution.messages cimport ModifyOrder
+from nautilus_trader.execution.messages cimport QueryAccount
+from nautilus_trader.execution.messages cimport QueryOrder
 from nautilus_trader.execution.messages cimport SubmitOrder
 from nautilus_trader.execution.messages cimport SubmitOrderList
 from nautilus_trader.model.events.account cimport AccountState
@@ -253,6 +255,22 @@ cdef class ExecutionClient(Component):
             f"You can implement by overriding the `batch_cancel_orders` method for this client",  # pragma: no cover  # noqa
         )
         raise NotImplementedError("method `batch_cancel_orders` must be implemented in the subclass")
+
+    cpdef void query_account(self, QueryAccount command):
+        """
+        Query the account specified by the command which will generate an `AccountState` event.
+
+        Parameters
+        ----------
+        command : QueryAccount
+            The command to execute.
+
+        """
+        self._log.error(  # pragma: no cover
+            f"Cannot execute command {command}: not implemented. "  # pragma: no cover
+            f"You can implement by overriding the `query_account` method for this client",  # pragma: no cover  # noqa
+        )
+        raise NotImplementedError("method `query_account` must be implemented in the subclass")
 
     cpdef void query_order(self, QueryOrder command):
         """